+++ plugins/modules/discovery.py
@@ -70,9 +70,10 @@ message:
     sample: 'Host created.'
 """
 
+import json
+
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.urls import fetch_url
-import json
 
 
 def run_module():
@@ -129,10 +130,14 @@ def run_module():
     )
 
     api_endpoint = (
-        "/objects/host/" + module.params.get("host_name") + "/actions/discover_services/invoke"
+        "/objects/host/"
+        + module.params.get("host_name")
+        + "/actions/discover_services/invoke"
     )
     url = base_url + api_endpoint
-    response, info = fetch_url(module, url, module.jsonify(params), headers=headers, method="POST")
+    response, info = fetch_url(
+        module, url, module.jsonify(params), headers=headers, method="POST"
+    )
     http_code = info["status"]
 
     # Kudos to Lars G.!