+++ plugins/modules/host.py
@@ -94,9 +94,10 @@ message:
     sample: 'Host created.'
 """
 
+import json
+
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.urls import fetch_url
-import json
 
 
 def exit_failed(module, msg):
@@ -142,8 +143,8 @@ def get_current_host_state(module, base_url, headers):
     else:
         exit_failed(
             module,
-            "Error calling API. HTTP code %d. Details: %s. Body: %s"
-            % (info["status"], info["body"], body),
+            "Error calling API. HTTP code %d. Details: %s."
+            % (info["status"], info.get("body", "N/A")),
         )
 
     return current_state, current_explicit_attributes, current_folder, etag
@@ -233,8 +234,8 @@ def run_module():
         automation_secret=dict(type="str", required=True, no_log=True),
         host_name=dict(type="str", required=True),
         attributes=dict(type="raw", default=[]),
-        folder=dict(type="str", default='/'),
-        state=dict(type="str", default='present', choices=["present", "absent"]),
+        folder=dict(type="str", default="/"),
+        state=dict(type="str", default="present", choices=["present", "absent"]),
     )
 
     module = AnsibleModule(argument_spec=module_args, supports_check_mode=False)
@@ -279,6 +280,9 @@ def run_module():
     if state == "present" and current_state == "present":
         headers["If-Match"] = etag
         msg_tokens = []
+
+        if current_folder.endswith("/"):
+            current_folder = current_folder.rstrip("/")
 
         if current_folder != module.params["folder"]:
             move_host(module, base_url, headers)